Emery's Audio Note Testing
Tested Emery's Audio Note Applicaiton in Unity: Audio Note is currently in Geology App, you can start the audio note with "begin note" or a press of the Audio note button. The dialogue recorded will print to a panel in the Unity game window and console. And save to a Geology text file in the Asset folder of the Unity File